engine driver abstraction
